DeSim: Simulating Derivatives
We present the first part of an article series on building the first version of an Options Simulator. The motivation was to evaluate a host of derivative products regarding their return and general performance. As a specific option type, we applied the simulator to everlasting options, introduced by Paradigm. This article first goes through the fundamental difference between perpetual futures and everlasting options to create the necessary background. Next, we cover how the options simulator was built and show its internal workings. Finally, we deep dive into everlasting options for crypto assets.